hamdullah
hamdullah

+++ database/migrations/2024_11_05_145654_add_role_to_users_table.php
@@ -9,14 +9,20 @@ class AddRoleToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('role')->default('user');
+            // Controleer of de kolom nog niet bestaat, zodat deze niet dubbel wordt aangemaakt
+            if (!Schema::hasColumn('users', 'role')) {
+                $table->string('role')->default('user');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            // Verwijder de kolom alleen als deze bestaat
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');
+            }
         });
     }
 }
